optimize log_like and log_prior
I tried to make the code a bit faster, especially log_like/calc_model
by using more list comprehension and the map function. I replace the
for loop in calc_model with a list comprehension and the for loop in
log_prior with map. This seems to have cut the execution time for
SEDBayesFitter.fit in half.