add ability to handle undetected fluxes
Undetected fluxes are plotted and downward facing arrows and are found
as NaNs in the “obs_flux” parameter. It is assumed the upper limit is
contained in the “obs_err” parameter. Model fluxes are only calculated
for the detected wavebands. Also adjusted the y-limits to account for
the model fluxes and the upper limits.